requirements.txt
python-gnupg
qrcode
pyzbar
PIL
